- Fix cavia .cads
- Fix some .cstr loops [Star Fox Assault (GC)]
- Add encrypted .scd [Final Fantasy XI (X360)]
- Add HCA key

+++ src/coding/coding.h
@@ -638,7 +638,7 @@ typedef struct {
     /* multichannel-only */
     int coupled_count;
     int stream_count;
-    int channel_mapping[255];
+    uint8_t channel_mapping[255];
     /* frame table */
     off_t table_offset;
     int table_count;
